fix wrapping

+++ src/py21cmfast/src/Stochasticity.h
@@ -44,7 +44,7 @@ int stochastic_halofield(unsigned long long int seed, float redshift_desc, float
                          HaloField *halos);
 
 int single_test_sample(unsigned long long int seed, int n_condition, float *conditions,
-                       int *cond_crd, double z_out, double z_in, int *out_n_tot, int *out_n_cell,
+                       float *cond_crd, double z_out, double z_in, int *out_n_tot, int *out_n_cell,
                        double *out_n_exp, double *out_m_cell, double *out_m_exp,
                        float *out_halo_masses, float *out_halo_pos);
