+++ src/py21cmfast/src/_functionprototypes_wrapper.h
@@ -105,10 +105,10 @@ void FunctionThatThrows();
 /*------------------------*/
 
 /* Test Outputs For Specific Models */
-int single_test_sample(int seed, int n_condition, float *conditions, int *cond_crd, double z_out,
-                       double z_in, int *out_n_tot, int *out_n_cell, double *out_n_exp,
-                       double *out_m_cell, double *out_m_exp, float *out_halo_masses,
-                       float *out_halo_pos);
+int single_test_sample(unsigned long long int seed, int n_condition, float *conditions,
+                       float *cond_crd, double z_out, double z_in, int *out_n_tot, int *out_n_cell,
+                       double *out_n_exp, double *out_m_cell, double *out_m_exp,
+                       float *out_halo_masses, float *out_halo_pos);
 // test function for getting halo properties from the wrapper, can use a lot of memory for large
 // catalogs
 int test_halo_props(double redshift, float *vcb_grid, float *J21_LW_grid, float *z_re_grid,
